brain_injury_simplify/requirements.txt
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.5.2.54
pandas==1.1.5
Pillow==8.4.0
scipy==1.5.4
tensorboardX==2.4.1
torch==1.9.0+cu111
torchsummary==1.5.1
torchvision==0.10.0+cu111
